# app/api/admin/audit-logs/route.ts
/**
 * Audit Log Viewer API
 *
 * GET /api/admin/audit-logs
 *
 * View audit logs with filtering and pagination.
 *
 * Security:
 * - Admin only
 * - Supports filtering by user, organization, event type, severity, date range
 * - Pagination for large result sets
 *
 * Sprint 9 - Story 9.9: Audit Logging & Monitoring
 */

import { NextRequest, NextResponse } from "next/server";
import { requirePermission, Permission } from "@/lib/rbac";
import {
  queryAuditLogs,
  getAuditLogStats,
  AuditEventType,
  AuditSeverity,
} from "@/lib/auditLog";

/**
 * GET /api/admin/audit-logs
 *
 * Query audit logs with filters.
 *
 * Query parameters:
 * - userId: Filter by user ID
 * - organizationId: Filter by organization ID
 * - eventType: Filter by event type (AUTH_LOGIN_SUCCESS, FILE_UPLOAD, etc.)
 * - severity: Filter by severity (INFO, WARNING, ERROR, CRITICAL)
 * - startDate: Filter by start date (ISO 8601)
 * - endDate: Filter by end date (ISO 8601)
 * - limit: Max results (default: 100, max: 1000)
 * - offset: Pagination offset
 * - format: Export format ('json' or 'csv')
 *
 * Returns: { logs: [], total: number, limit: number, offset: number }
 * Or CSV file if format=csv
 */
export async function GET(request: NextRequest) {
  try {
    // Require admin permission
    await requirePermission(Permission.VIEW_AUDIT_LOGS);

    const { searchParams } = new URL(request.url);

    // Parse filters
    const userId = searchParams.get("userId") || undefined;
    const organizationId = searchParams.get("organizationId") || undefined;
    const eventType = searchParams.get("eventType") as
      | AuditEventType
      | undefined;
    const severity = searchParams.get("severity") as AuditSeverity | undefined;
    const startDateStr = searchParams.get("startDate");
    const endDateStr = searchParams.get("endDate");
    const limit = parseInt(searchParams.get("limit") || "100", 10);
    const offset = parseInt(searchParams.get("offset") || "0", 10);
    const format = searchParams.get("format") || "json";

    // Parse dates
    const startDate = startDateStr ? new Date(startDateStr) : undefined;
    const endDate = endDateStr ? new Date(endDateStr) : undefined;

    // Validate event type if provided
    if (eventType && !Object.values(AuditEventType).includes(eventType)) {
      return NextResponse.json(
        { error: "Invalid event type" },
        { status: 400 }
      );
    }

    // Validate severity if provided
    if (severity && !Object.values(AuditSeverity).includes(severity)) {
      return NextResponse.json(
        { error: "Invalid severity level" },
        { status: 400 }
      );
    }

    // Query audit logs
    const result = await queryAuditLogs({
      userId,
      organizationId,
      eventType,
      severity,
      startDate,
      endDate,
      limit,
      offset,
    });

    // Export as CSV if requested
    if (format === "csv") {
      const csv = convertLogsToCSV(result.logs);
      const filename = `audit-logs-${new Date().toISOString().split("T")[0]}.csv`;

      return new NextResponse(csv, {
        status: 200,
        headers: {
          "Content-Type": "text/csv",
          "Content-Disposition": `attachment; filename="${filename}"`,
          "Cache-Control": "no-cache",
        },
      });
    }

    return NextResponse.json(result);
    // FIX: Use unknown type with type guard
  } catch (error: unknown) {
    console.error("Error querying audit logs:", error);

    return NextResponse.json(
      { error: "Failed to query audit logs" },
      { status: 500 }
    );
  }
}

// Type for audit log entries
interface AuditLogEntry {
  createdAt?: string | Date;
  eventType?: string;
  severity?: string;
  userId?: string | null;
  organizationId?: string | null;
  ipAddress?: string | null;
  userAgent?: string | null;
  resourceType?: string | null;
  resourceId?: string | null;
  action?: string | null;
  details?: unknown;
  [key: string]: unknown;
}

/**
 * Convert audit logs to CSV format
 */
function convertLogsToCSV(logs: AuditLogEntry[]): string {
  // CSV header
  const header = [
    "Timestamp",
    "Event Type",
    "Severity",
    "User ID",
    "Organization ID",
    "IP Address",
    "User Agent",
    "Resource Type",
    "Resource ID",
    "Action",
    "Details",
  ].join(",");

  // CSV rows
  const rows = logs.map((log) => {
    const row = [
      log.createdAt || "",
      log.eventType || "",
      log.severity || "",
      log.userId || "",
      log.organizationId || "",
      log.ipAddress || "",
      escapeCSV(log.userAgent || ""),
      log.resourceType || "",
      log.resourceId || "",
      log.action || "",
      escapeCSV(JSON.stringify(log.details || {})),
    ];

    return row.map(String).join(",");
  });

  return [header, ...rows].join("\n");
}

/**
 * Escape CSV values to handle commas, quotes, and newlines
 */
function escapeCSV(value: string): string {
  if (value.includes(",") || value.includes('"') || value.includes("\n")) {
    return `"${value.replace(/"/g, '""')}"`;
  }
  return value;
}

/**
 * GET /api/admin/audit-logs/stats
 *
 * Get audit log statistics.
 *
 * Query parameters:
 * - organizationId: Filter by organization
 * - startDate: Filter by start date
 * - endDate: Filter by end date
 *
 * Returns statistics about audit log events.
 */
export async function stats(request: NextRequest) {
  try {
    await requirePermission(Permission.VIEW_AUDIT_LOGS);

    const { searchParams } = new URL(request.url);
    const organizationId = searchParams.get("organizationId") || undefined;
    const startDateStr = searchParams.get("startDate");
    const endDateStr = searchParams.get("endDate");

    const startDate = startDateStr ? new Date(startDateStr) : undefined;
    const endDate = endDateStr ? new Date(endDateStr) : undefined;

    const stats = await getAuditLogStats(organizationId, startDate, endDate);

    return NextResponse.json(stats);
    // FIX: Use unknown type with type guard
  } catch (error: unknown) {
    console.error("Error getting audit log stats:", error);

    return NextResponse.json(
      { error: "Failed to get audit log statistics" },
      { status: 500 }
    );
  }
}