# app/api/admin/users/route.ts
import { NextRequest, NextResponse } from "next/server";
import { db } from "@/lib/db";
import { Prisma, UserRole } from "@prisma/client";
import { requirePermission, Permission } from "@/lib/rbac";
import { z } from "zod";
import { zodErrorResponse } from "@/lib/validation";
// M1 FIX: Add CSRF validation
import { validateCSRFWithMobile } from "@/lib/csrf";

// GET /api/admin/users - List all users
export async function GET(request: NextRequest) {
  try {
    await requirePermission(Permission.VIEW_USERS);

    const { searchParams } = request.nextUrl;
    const page = parseInt(searchParams.get("page") || "1");
    const limit = parseInt(searchParams.get("limit") || "20");
    const role = searchParams.get("role");
    const search = searchParams.get("search");

    // H1 FIX: Use typed Prisma where input instead of any
    const where: Prisma.UserWhereInput = {};

    // H14 FIX: Validate role against UserRole enum
    if (role && Object.values(UserRole).includes(role as UserRole)) {
      where.role = role as UserRole;
    }

    if (search) {
      where.OR = [
        { email: { contains: search, mode: "insensitive" } },
        { firstName: { contains: search, mode: "insensitive" } },
        { lastName: { contains: search, mode: "insensitive" } },
      ];
    }

    const [users, total] = await Promise.all([
      db.user.findMany({
        where,
        select: {
          id: true,
          email: true,
          phone: true,
          firstName: true,
          lastName: true,
          role: true,
          isActive: true,
          isEmailVerified: true,
          isPhoneVerified: true,
          organizationId: true,
          organization: {
            select: {
              id: true,
              name: true,
              type: true,
              isVerified: true,
            },
          },
          createdAt: true,
          lastLoginAt: true,
        },
        skip: (page - 1) * limit,
        take: limit,
        orderBy: {
          createdAt: "desc",
        },
      }),
      db.user.count({ where }),
    ]);

    return NextResponse.json({
      users,
      pagination: {
        page,
        limit,
        total,
        pages: Math.ceil(total / limit),
      },
    });
  } catch (error) {
    // L6 FIX: Server-side logging intentional for debugging - errors not exposed to client
    console.error("List users error:", error);

    if (error instanceof Error && error.name === "ForbiddenError") {
      return NextResponse.json({ error: error.message }, { status: 403 });
    }

    return NextResponse.json(
      { error: "Internal server error" },
      { status: 500 }
    );
  }
}

const updateUserSchema = z.object({
  userId: z.string(),
  role: z.enum(["SHIPPER", "CARRIER", "DISPATCHER", "ADMIN", "SUPER_ADMIN"]),
});

// PATCH /api/admin/users - Update user role
export async function PATCH(request: NextRequest) {
  try {
    // M1 FIX: Add CSRF validation
    const csrfError = await validateCSRFWithMobile(request);
    if (csrfError) return csrfError;

    await requirePermission(Permission.ASSIGN_ROLES);

    const body = await request.json();
    const { userId, role } = updateUserSchema.parse(body);

    const user = await db.user.update({
      where: { id: userId },
      data: { role },
      select: {
        id: true,
        email: true,
        firstName: true,
        lastName: true,
        role: true,
      },
    });

    return NextResponse.json({
      message: "User role updated successfully",
      user,
    });
  } catch (error) {
    // L6 FIX: Server-side logging intentional for debugging - errors not exposed to client
    console.error("Update user role error:", error);

    if (error instanceof Error && error.name === "ForbiddenError") {
      return NextResponse.json({ error: error.message }, { status: 403 });
    }

    if (error instanceof z.ZodError) {
      return zodErrorResponse(error);
    }

    return NextResponse.json(
      { error: "Internal server error" },
      { status: 500 }
    );
  }
}