# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freight Management Platform API
  description: |
    Comprehensive API for Ethiopian freight logistics platform.

    ## Authentication
    Most endpoints require authentication via session cookie or Bearer token.
    - **Web clients**: Use session cookie with CSRF token in X-CSRF-Token header
    - **Mobile clients**: Use Bearer token in Authorization header

    ## Rate Limiting
    All endpoints are rate-limited. When exceeded, returns 429 with Retry-After header.

    ## Error Handling
    - 400: Validation error (Zod schema)
    - 401: Unauthorized
    - 403: Forbidden (insufficient permissions)
    - 404: Resource not found
    - 409: Conflict (state machine violation)
    - 429: Rate limit exceeded
    - 500: Internal server error
  version: 1.0.0
  contact:
    name: Freight Platform Support
    email: support@freightplatform.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: https://api.freightplatform.com/api
    description: Production
...
paths:
...
  /escalations/{id}:
    patch:
      tags: [dispatcher]
      summary: Resolve escalation
      operationId: resolveEscalation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Escalation resolved

  # ============================================================================
  # ADMIN PANEL
  # ============================================================================
...
  /admin/users:
    get:
      tags: [admin]
      summary: List all users
      operationId: listUsers
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/UserRole"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    patch:
      tags: [admin]
      summary: Update user role
      operationId: updateUserRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, role]
              properties:
                userId:
                  type: string
                role:
                  $ref: "#/components/schemas/UserRole"
      responses:
        "200":
          description: Role updated
...
  /admin/audit-logs:
    get:
      tags: [admin]
      summary: List audit logs
      operationId: listAuditLogs
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: action
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit log entries
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    UserRole:
      type: string
      enum: [SHIPPER, CARRIER, DISPATCHER, ADMIN, SUPER_ADMIN]
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          type: string
        organizationId:
          type: string
        isEmailVerified:
          type: boolean
        isPhoneVerified:
          type: boolean
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            organization:
              $ref: "#/components/schemas/OrganizationSummary"
            createdAt:
              type: string
              format: date-time
            lastLoginAt:
              type: string
              format: date-time
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # --------------------------------------------------------------------------
    # ORGANIZATION
    # --------------------------------------------------------------------------
    OrganizationSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        isVerified:
          type: boolean

    # --------------------------------------------------------------------------
    # LOAD
    # --------------------------------------------------------------------------
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
security:
  - bearerAuth: []
  - cookieAuth: []